Avalon

+++ Avalon/Avalon.Network/Case.cs
@@ -2,5 +2,5 @@ namespace Avalon.Network;
 
 public class Case : Any
 {
-    public virtual int Index { get; set; }
+    public virtual long Index { get; set; }
 }
